move and split files
