+++ src/const.py
@@ -8,12 +8,12 @@ SCREEN_HEIGHT = 50
 MAP_WIDTH = 80
 MAP_HEIGHT = 43
 
-FONT_PATH = 'data/cp437_12x12.png'
+FONT_PATH = 'data/font/cp437_12x12.png'
 FONT_FLAG = tcod.FONT_TYPE_GREYSCALE | tcod.FONT_LAYOUT_ASCII_INROW
 
 FOV_ALGORITHM = tcod.FOV_BASIC
 FOV_LIGHT_WALLS = True
-FOV_RADIUS = 10
+FOV_RADIUS = 80  # 10
 
 
 COLORS = {
