+++ src/input_handler.py
@@ -1,5 +1,4 @@
 from dataclasses import dataclass
-
 import tcod
